refactor(kalendar): уклони feasts.py, fix_moveable_feasts, migracija_slava (#259)
JSONL је сада једини извор слава, а рачун датума покретних празника је у
моделу (Slava.get_datum/calc_vaskrs) — па посебан модул и команде за
конверзију више нису потребни.

- обрисан crkva/kalendar/feasts.py (подаци су у fixtures/slave.jsonl,
  offset_dani по празнику; рачун у моделу).
- обрисан fix_moveable_feasts.py — на свежем увозу непотребан (slave.jsonl
  већ носи покретне као покретне, без фиксних копија); проверено пуним
  rebuild-ом (Чукарица из crkva.zip + mock Амстердам).
- обрисан migracija_slava.py — слава се више не увози из DBF-а (био извор
  дупликата #259), већ из slave.jsonl преко unos_slava.
- importuj_dbf: слава корак сада зове unos_slava (slave.jsonl) уместо
  migracija_slava; уклоњен fix_moveable_feasts.
- load_dbf: уклоњен HSPSLAVE.DBF мапинг (hsp_slave staging без потрошача).
- тестови пресложени у kalendar/tests/test_slave_jsonl.py (структура JSONL,
  unos_slava, рачун покретних датума у моделу).
